Extracted Walker's parser and related classes into `joe.parser`.
