+++ lib/src/main/java/com/wjduquette/joe/parser/Expr.java
@@ -1,4 +1,4 @@
-package com.wjduquette.joe.walker;
+package com.wjduquette.joe.parser;
 
 import com.wjduquette.joe.scanner.Token;
 import java.util.List;
