+++ lib/src/main/java/com/wjduquette/joe/walker/Resolver.java
@@ -2,6 +2,8 @@ package com.wjduquette.joe.walker;
 
 import com.wjduquette.joe.JoeClass;
 import com.wjduquette.joe.Trace;
+import com.wjduquette.joe.parser.Expr;
+import com.wjduquette.joe.parser.Stmt;
 import com.wjduquette.joe.scanner.Token;
 
 import java.util.HashMap;
