+++ lib/src/test/java/com/wjduquette/joe/walker/ParserTest.java
@@ -1,5 +1,6 @@
 package com.wjduquette.joe.walker;
 
+import com.wjduquette.joe.parser.Parser;
 import com.wjduquette.joe.scanner.SourceBuffer;
 import com.wjduquette.joe.Ted;
 import org.junit.Test;
